Document Docker flag needed on Linux (#41)
When you are running Docker on Linux you need to add `--add-host host.docker.internal:host-gateway` to the flags of `docker run` to be able to connect to the host as explained in [this StackOverflow answer](https://stackoverflow.com/a/24326540/2493714). We need to be able to connect to the host so we can use the `esp_rfc2217_server.py` running on the host.